Pod Handlers and Refactor
refactor: template_name -> name

fix: add method to logging

fix: change some `logger.error` to `logger.exception`

fix: use adaptive task creating

refactor: remove unused get_k8s_api method of Repo

fix: fix shortuuid illegal alphabet

refactor: remove Admin prefix of services

feat: finish pod handles

feat: add heartbeat service

feat: add k8s configs, fmts

refactor: parse_template_str -> render_template_str

refactor: merge create and update events

fix: fix template

docs: update README.md

+++ src/apiserver/service/handler.py
@@ -1,4 +1,7 @@
+import asyncio
 import base64
+import datetime
+import json
 from typing import Optional, Union
 
 import kubernetes
@@ -7,15 +10,17 @@ from kubernetes.client import ApiException
 from loguru import logger
 from pydantic import BaseModel
 
-from src.components import config
+from src.components import config, errors
 from src.components.config import CONFIG_K8S_CREDENTIAL_FMT
+from src.components.datamodels import PodStatusEnum, UserStatusEnum, ResourceStatusEnum
 from src.components.events import (
     TemplateCreateEvent, TemplateUpdateEvent, TemplateDeleteEvent,
     UserCreateEvent, UserUpdateEvent, UserDeleteEvent,
-    PodCreateEvent, PodUpdateEvent, PodDeleteEvent, PodTimeoutEvent,
+    PodCreateUpdateEvent, PodDeleteEvent, PodTimeoutEvent,
     UserHeartbeatEvent
 )
 import src.apiserver.service
+from src.components.utils import render_template_str
 
 
 async def handle_template_create_event(srv: Optional['src.apiserver.service.RootService'],
@@ -56,31 +61,15 @@ async def handle_user_create_event(srv: Optional['src.apiserver.service.RootServ
         return None
     if user is not None:
         logger.info(f"creating k8s credentials for user {ev.username}")
-        err = None
-        try:
-            ret = srv.k8s_operator_service.v1.create_namespaced_secret(
-                config.CONFIG_PROJECT_NAMESPACE,
-                kubernetes.client.V1Secret(
-                    api_version="v1",
-                    kind="Secret",
-                    data={
-                        "auth": base64.b64encode(user.htpasswd.get_secret_value().encode()).decode()
-                    },
-                    metadata=kubernetes.client.V1ObjectMeta(
-                        name=CONFIG_K8S_CREDENTIAL_FMT.format(user.username),
-                        namespace=config.CONFIG_PROJECT_NAMESPACE
-                    )
-                ))
-            if ret is None:
-                err = RuntimeError("failed to create k8s secret")
-        except ApiException as e:
-            if e.reason == 'Conflict':
-                logger.warning(f"secret {CONFIG_K8S_CREDENTIAL_FMT.format(user.username)} already exists")
-            else:
-                logger.error(e)
-                err = e
-        if err is None:
-            return RuntimeError("failed to create k8s secret")
+
+        err = await srv.k8s_operator_service.create_or_update_user_credentials(
+            user.username,
+            user.htpasswd.get_secret_value().encode()
+        )
+
+        if err is not None:
+            logger.error(f"handle_user_create_event failed to create k8s credentials: {err}")
+            return err
         else:
             err = await srv.user_service.repo.commit(ev.username)
             if err is not None:
@@ -99,32 +88,13 @@ async def handle_user_update_event(srv: Optional['src.apiserver.service.RootServ
         return None
     if user is not None:
         logger.info(f"updating k8s credentials for user {ev.username}")
-        err = None
-        try:
-            ret = srv.k8s_operator_service.v1.patch_namespaced_secret(
-                CONFIG_K8S_CREDENTIAL_FMT.format(user.username),
-                config.CONFIG_PROJECT_NAMESPACE,
-                kubernetes.client.V1Secret(
-                    api_version="v1",
-                    kind="Secret",
-                    data={
-                        "auth": base64.b64encode(user.htpasswd.get_secret_value().encode()).decode()
-                    },
-                    metadata=kubernetes.client.V1ObjectMeta(
-                        name=CONFIG_K8S_CREDENTIAL_FMT.format(user.username),
-                        namespace=config.CONFIG_PROJECT_NAMESPACE
-                    )
-                ))
-            if ret is None:
-                err = RuntimeError("failed to update k8s secret")
-        except ApiException as e:
-            if e.reason == 'Not Found':
-                logger.warning(f"secret {CONFIG_K8S_CREDENTIAL_FMT.format(user.username)}not found")
-            else:
-                logger.error(e)
-                err = e
-        if err is None:
-            return RuntimeError("failed to update k8s secret")
+        err = await srv.k8s_operator_service.create_or_update_user_credentials(
+            user.username,
+            user.htpasswd.get_secret_value().encode()
+        )
+        if err is not None:
+            logger.error(f"handle_user_update_event failed to create k8s credentials: {err}")
+            return err
         else:
             err = await srv.user_service.repo.commit(ev.username)
             if err is not None:
@@ -142,53 +112,103 @@ async def handle_user_delete_event(srv: Optional['src.apiserver.service.RootServ
         logger.warning(err)
         return None
     if user is not None:
-        # Delete Credentials
+        # delete credentials
         logger.info(f"deleting k8s credentials for user {ev.username}")
-        err = None
-        try:
-            ret = srv.k8s_operator_service.v1.delete_namespaced_secret(
-                CONFIG_K8S_CREDENTIAL_FMT.format(user.username),
-                config.CONFIG_PROJECT_NAMESPACE,
-            )
-            if ret is None:
-                err = RuntimeError("failed to delete k8s secret")
-        except ApiException as e:
-            if e.reason == 'Not Found':
-                logger.warning(f"secret {CONFIG_K8S_CREDENTIAL_FMT.format(user.username)} not found")
-            else:
-                logger.error(e)
-                err = e
-
-        # Delete Pods
-
+        err = await srv.k8s_operator_service.delete_user_credential(
+            user.username,
+        )
         if err is not None:
+            logger.error(f"handle_user_delete_event failed to delete k8s credentials: {err}")
+            return err
+
+        # delete pods
+        logger.info(f"deleting pods for user {ev.username}")
+        _, pods, err = await srv.pod_service.repo.list(extra_query_filter={"username": ev.username})
+        if err is not None:
+            logger.error(f"handle_user_delete_event failed to list pods: {err}")
+            return err
+        for pod in pods:
+            if pod.username != ev.username:
+                logger.error(f"handle_user_delete_event found pod {pod.pod_id} with wrong username {pod.username}")
+
+            _, err = await srv.pod_service.repo.delete(pod_id=pod.pod_id)
+            if err is not None:
+                logger.error(f"handle_user_delete_event failed to delete pod {pod.pod_id}: {err}")
+                return err
+
+            err = await handle_pod_delete_event(srv, PodDeleteEvent(pod_id=pod.pod_id, username=pod.username))
+            if err is not None:
+                logger.error(f"handle_user_delete_event failed to delete pod {pod.pod_id}: {err}")
+                return err
+
+            _, err = await srv.pod_service.repo.purge(pod_id=pod.pod_id)
+            if err is not None:
+                logger.error(f"handle_user_delete_event failed to purge pod {pod.pod_id}: {err}")
+                return err
+
+        # finally, purge user
+        _, err = await srv.user_service.repo.purge(ev.username)
+        if err is not None:
+            logger.error(f"handle_user_delete_event failed to commit: {err}")
             return err
         else:
-            _, err = await srv.user_service.repo.purge(ev.username)
-            if err is not None:
-                logger.error(f"handle_user_delete_event failed to commit: {err}")
-                return err
-            else:
-                return err
+            return err
 
 
-async def handle_pod_create_event(srv: Optional['src.apiserver.service.RootService'],
-                                  ev: Union[PodCreateEvent, BaseModel]) -> Optional[Exception]:
-    # TODO: create pod resources in namespace
-    err = await srv.pod_service.repo.commit(ev.pod_id)
-    if err is not None:
-        logger.error(f"handle_pod_create_event failed to commit: {err}")
-        return err
-    else:
+async def handle_pod_create_update_event(srv: Optional['src.apiserver.service.RootService'],
+                                         ev: Union[PodCreateUpdateEvent, BaseModel]) -> Optional[Exception]:
+    # check user's validity
+    user, err = await srv.user_service.repo.get(ev.username)
+    if any([
+        err is not None,
+        user.status != UserStatusEnum.active,
+        user.resource_status == ResourceStatusEnum.deleted,
+    ]):
+        logger.error(f"handle_pod_create_update_event failed to get user {ev.username}: {err}")
         return err
 
+    # check if pod is matched
+    pod, err = await srv.pod_service.repo.get(ev.pod_id)
+    if any([
+        err is not None,
+        pod.resource_status == ResourceStatusEnum.deleted,
+    ]):
+        logger.error(f"handle_pod_create_update_event failed to get pod {ev.pod_id}: {err}")
+        return err
 
-async def handle_pod_update_event(srv: Optional['src.apiserver.service.RootService'],
-                                  ev: Union[PodUpdateEvent, BaseModel]) -> Optional[Exception]:
-    # TODO: update pod resources in namespace
+    # get template
+    template, err = await srv.template_service.repo.get(str(pod.template_ref))
+    if err is not None:
+        logger.error(f"handle_pod_create_update_event failed to get template {pod.template_ref}: {err}")
+        return err
+
+    # render template
+    kv = pod.values | srv.opt.k8s_config_values | template.values
+    template_str, _, err = render_template_str(template.template_str, kv)
+    if err is not None:
+        logger.error(f"handle_pod_create_update_event failed to parse template {pod.template_ref}: {err}")
+        return err
+
+    # operate on k8s
+    err = await srv.k8s_operator_service.create_or_update_pod(pod.pod_id, template_str)
+    if err is not None:
+        logger.error(f"handle_pod_create_update_event failed to create pod {pod.pod_id}: {err}")
+        return err
+
+    # update pod's status
+    _, err = await srv.pod_service.repo.update(
+        pod_id=pod.pod_id,
+        started_at=datetime.datetime.now(),
+        current_status=pod.target_status,  # FIXME: might cause trouble
+    )
+    if err is not None:
+        logger.error(f"handle_pod_create_update_event failed to update pod {pod.pod_id}: {err}")
+        return err
+
+    # finally commit
     err = await srv.pod_service.repo.commit(ev.pod_id)
     if err is not None:
-        logger.error(f"handle_pod_update_event failed to commit: {err}")
+        logger.error(f"handle_pod_create_update_event failed to commit: {err}")
         return err
     else:
         return err
@@ -196,7 +216,11 @@ async def handle_pod_update_event(srv: Optional['src.apiserver.service.RootServi
 
 async def handle_pod_delete_event(srv: Optional['src.apiserver.service.RootService'],
                                   ev: Union[PodDeleteEvent, BaseModel]) -> Optional[Exception]:
-    # TODO: delete pod resources in namespace
+    err = await srv.k8s_operator_service.delete_pod(ev.pod_id)
+    if err is not None:
+        logger.error(f"handle_pod_delete_event failed to delete pod {ev.pod_id}: {err}")
+        return err
+
     _, err = await srv.pod_service.repo.purge(ev.pod_id)
     if err is not None:
         logger.error(f"handle_pod_delete_event failed to commit: {err}")
@@ -207,11 +231,12 @@ async def handle_pod_delete_event(srv: Optional['src.apiserver.service.RootServi
 
 async def handle_user_heartbeat_event(srv: Optional['src.apiserver.service.RootService'],
                                       ev: Union[UserHeartbeatEvent, BaseModel]) -> Optional[Exception]:
-    # TODO: update user heartbeat
-    pass
+    err = await srv.heartbeat_service.ping(ev.username)
+    if err is not None:
+        logger.error(f"handle_user_heartbeat_event failed to list pods: {err}")
+        return err
 
-
-async def handle_pod_timeout_event(srv: Optional['src.apiserver.service.RootService'],
-                                   ev: Union[PodTimeoutEvent, BaseModel]) -> Optional[Exception]:
-    # TODO: handle pod timeout
-    pass
+# async def handle_pod_timeout_event(srv: Optional['src.apiserver.service.RootService'],
+#                                    ev: Union[PodTimeoutEvent, BaseModel]) -> Optional[Exception]:
+#     # TODO: handle pod timeout
+#     pass
